+++ packages/zosuss/src/SshSession.ts
@@ -55,6 +55,11 @@ export class SshSession {
     };
 
     /**
+     * @deprecated Use SSH_OPTION_HOST
+     */
+    public static SSH_OPTION_HOST_PROFILE: ICommandOptionDefinition = this.SSH_OPTION_HOST;
+
+    /**
      * Option used in profile creation and commands for port for z/OS SSH
      */
     public static SSH_OPTION_PORT: ICommandOptionDefinition = {
@@ -77,6 +82,11 @@ export class SshSession {
         required: false,
         group: SshSession.SSH_CONNECTION_OPTION_GROUP
     };
+
+    /**
+     * @deprecated Use SSH_OPTION_USER
+     */
+    public static SSH_OPTION_USER_PROFILE: ICommandOptionDefinition = this.SSH_OPTION_USER;
 
     /**
      * Option used in profile creation and commands for password/passphrase for z/OS SSH